create TitleGridContainer component and integrate into skill and experience section, extract data from skill and experience into other files
