+++ src/assets/data/skillsData.js
@@ -1,0 +1,55 @@
+const programmingLangs = {
+    title: 'Programming languages',
+    content: [
+    'JavaScript',
+    'Swift',
+    'Objective-C',
+    'HTML',
+    'CSS',
+    'C#',
+    'Java',
+    'C++'
+    ]
+};
+
+const frameworks = {
+    title: 'Frameworks',
+    content: [
+    'Node.js',
+    'React.js',
+    'Redux.js',
+    'Express.js',
+    'IOS MVC',
+    'CodeIgniter'
+    ]
+};
+
+const packageManagements = {
+    title: 'Package Management',
+    content: [
+    'NPM',
+    'CocoaPods',
+    ]
+};
+
+const versionControls = {
+    title: 'Version Control',
+    content: [
+    'Git',
+    'GitHub',
+    'Bitbucket'
+    ]
+};
+
+export const skillList = [
+    programmingLangs,
+    frameworks,
+    packageManagements,
+    versionControls
+]
+
+export const skillDescription = `An artist of considerable range, Chet Faker — the name taken by
+Melbourne-raised, Brooklyn-based Nick Murphy — writes, performs
+and records all of his own music, giving it a warm, intimate
+feel with a solid groove structure.`;
+
