+++ src/assets/jss/material-kit-react/components/titleGridStyle.js
@@ -1,0 +1,17 @@
+const titleGridStyle = {
+    title: {
+        margin: '1%',
+        padding: '1%',
+        fontSize: '2rem',
+        borderBottom: '1px solid black',
+        textAlign: 'center !important'
+    },
+    description: {
+        fontWeight: '400',
+        fontSize: '1rem',
+        color: "#000000",
+        textAlign: "center !important"
+    },
+}
+
+export default titleGridStyle;
