+++ src/assets/jss/material-kit-react/sections/skillsStyle.js
@@ -1,19 +1,4 @@
 const skillStyle = {
-    title: {
-        margin: '1%',
-        padding: '1%',
-        fontSize: '2rem',
-        borderBottom: '1px solid black',
-        textAlign: 'center !important'
-    },
-    description: {
-        // margin: "1.071rem auto 0",
-        // maxWidth: "600px",
-        fontWeight: '400',
-        fontSize: '1rem',
-        color: "#000000",
-        textAlign: "center !important"
-    },
     panelTitle: {
         fontWeight: '600'
     },
