+++ src/views/Sections/SectionExperience.js
@@ -8,52 +8,17 @@ import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 
 //import from template componets
-import GridContainer from "components/Grid/GridContainer.js";
-import GridItem from 'components/Grid/GridItem.js';
+import TitleGridContainer from 'views/Components/TitleGridContainer';
 
 //style and theme
 import experienceStyle from 'assets/jss/material-kit-react/sections/experienceStyle.js';
-import skillTheme from 'assets/theme/skillsTheme.js';
+import panelTheme from 'assets/theme/expansionPanelTheme.js';
 
 //react-markdown
 import ReactMarkdown from 'react-markdown/with-html';
 
-//work exp markdown
-import jtiotMD from 'assets/markdownSrc/experience/jtiot.md';
-import artechMD from 'assets/markdownSrc/experience/artech.md'
-import luckyMD from 'assets/markdownSrc/experience/lucky.md';
-import weiQiaoMD from 'assets/markdownSrc/experience/wei-qiao.md';
-import mobiliStudioMD from 'assets/markdownSrc/experience/mobili-studio.md'
-
-const breakPoints = {
-    xs: 11,
-    sm: 11,
-    md: 8
-}
-
 //experience data
-const experiences = [
-    {
-        title: 'JTIOT',
-        mdSrc: jtiotMD
-    },
-    {
-        title: 'Artech',
-        mdSrc: artechMD
-    },
-    {
-        title: 'Lucky',
-        mdSrc: luckyMD
-    },
-    {
-        title: 'Wei Qiao',
-        mdSrc: weiQiaoMD
-    },
-    {
-        title: 'Mobili Studio',
-        mdSrc: mobiliStudioMD
-    }
-]
+import {getExperiences, expDescription} from 'assets/data/expData';
 
 class SectionExperience extends React.Component {
     constructor(){
@@ -65,22 +30,9 @@ class SectionExperience extends React.Component {
     }
 
     componentWillMount(){
-        const mdSrcs = experiences.map(exp=>{
-            return fetch(exp.mdSrc)
-            .then(res=>res.text())
-            .catch(err=>console.log(err));
-        });
 
-        Promise.all(mdSrcs)
-        .then(res=>{
-            const mds = res.map((mdText, index)=>{
-                return {
-                    title: experiences[index].title,
-                    content: mdText
-                }
-            })
-            this.setState({...this.state, md:mds});
-        });
+        getExperiences()
+        .then(res=>this.setState({...this.state, md:res}));
     }
 
     handlePanelExpand(panelName){
@@ -94,51 +46,43 @@ class SectionExperience extends React.Component {
 
     renderExpMD(exps){
         const {classes} = this.props;
-        return exps.map((exp, index)=>{
-            return (
-                <ExpansionPanel
-                key={index}
-                expanded={this.state.expand === `panel${index+1}`} 
-                onChange={()=>this.handlePanelExpand(`panel${index+1}`)}>
-                    <ExpansionPanelSummary
-                    expandIcon={<ExpandMoreIcon />}
-                    >
-                        <div className={classes.panelTitle}>{exp.title}</div>
-                    </ExpansionPanelSummary>
-                    <ExpansionPanelDetails>
-                        <div>
-                            <ReactMarkdown source={exp.content} escapeHtml={false} />
-                        </div>
-                    </ExpansionPanelDetails>
-                </ExpansionPanel>
-            );
-        })
+        if(!exps) return null;
+        return(
+            <ThemeProvider theme={panelTheme}>
+                <div>
+                {
+                    exps.map((exp, index)=>{
+                        return (
+                            <ExpansionPanel
+                            key={index}
+                            expanded={this.state.expand === `panel${index+1}`} 
+                            onChange={()=>this.handlePanelExpand(`panel${index+1}`)}>
+                                <ExpansionPanelSummary
+                                expandIcon={<ExpandMoreIcon />}
+                                >
+                                    <div className={classes.panelTitle}>{exp.title}</div>
+                                </ExpansionPanelSummary>
+                                <ExpansionPanelDetails>
+                                    <div>
+                                        <ReactMarkdown source={exp.content} escapeHtml={false} />
+                                    </div>
+                                </ExpansionPanelDetails>
+                            </ExpansionPanel>
+                        );
+                    })
+                }
+                </div>
+            </ThemeProvider>
+        )
     }
 
     render(){
-        const {classes} = this.props;
         return (
-            <ThemeProvider theme={skillTheme}>
-            <GridContainer direction='column' justify="center" alignItems='center'>
-                <GridItem {...breakPoints}>
-                    <div className={classes.title}>Experience</div>
-                </GridItem>
-                <GridItem {...breakPoints}>
-                    <div>
-                        <p className={classes.description}>
-                        Experience description here 
-                        </p>
-                    </div>
-                </GridItem>
-                <GridItem {...breakPoints}>
-                    {
-                        this.state.md?
-                        this.renderExpMD(this.state.md)
-                        :null
-                    }
-                </GridItem>
-            </GridContainer>
-            </ThemeProvider>
+            <TitleGridContainer title='Experiences' desc={expDescription}>
+            {
+                this.renderExpMD(this.state.md)
+            }
+            </TitleGridContainer>
         );
     }
